fix: almanax and party channel custom config
