+++ src/config.js
@@ -17,15 +17,15 @@ export default {
     '🌈': 'Hup',
     '💢': 'Sac',
     '😼': 'Eca',
-    '🐶': 'Kilorf',
+    '🐶': 'Ouginak',
     '💀': 'Sram',
     '🥊': 'Iop',
     '🏹': 'Cra',
     '🌱': 'Sadi',
     '🐼': 'Panda',
-    '💣': 'Ladino',
-    '🎭': 'Zobal',
-    '🤖': 'Steamer',
+    '💣': 'Rogue',
+    '🎭': 'Mask',
+    '🤖': 'Fogger',
     '🌀': 'Elio'
   },
   rarityMap: {
