+++ src/models/guild.js
@@ -7,11 +7,16 @@ const GuildSchema = new mongoose.Schema({
   },
   lang: {
     type: String,
-    default: 'en',
     validate: {
       validator: (lang) => ['en', 'pt', 'fr', 'es'].some((validLang) => validLang === lang),
       message: props => `${props.value} is not a valid language`
     }
+  },
+  almanaxChannel: {
+    type: String
+  },
+  partyChannel: {
+    type: String
   }
 })
